+++ python/GafferSceneTest/OpenGLRenderTest.py
@@ -114,6 +114,11 @@ class OpenGLRenderTest( GafferSceneTest.RenderTest ) :
 
 		pass
 
+	@unittest.skip( "Light linking not supported" )
+	def testLightLinkingWithExclusions( self ) :
+
+		pass
+
 	@unittest.skip( "Shadow linking not supported" )
 	def testShadowLinking( self ) :